requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==78.1.0
ttkthemes==3.2.2
urllib3==2.4.0
